pyproject.toml
[tool.poetry]
name = "alfred-slack"
version = "0.1.0"
description = ""
authors = ["TumuGuskun <johntgaskin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
slack-sdk = "^3.17.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
